ingestion_layer/requirements_ingestion.txt
apache-airflow
kafka-python
pandas
faker